refactor: api 하위 폴더 index 파일로 변경
